Fix no-use-before-define lint error in tab-bar.tsx
