+++ app/src/ui/tab-bar-type.ts
@@ -1,0 +1,11 @@
+/** The tab bar type. */
+export enum TabBarType {
+  /** Standard tabs */
+  Tabs,
+
+  /** Simpler switch appearance */
+  Switch,
+
+  /** Vertical tabs */
+  Vertical,
+}
